refactor(api)!: retire fx.run_metrics surface (#448)
run_metrics, run_metrics_iter, run_metrics_chunked, MetricsBundle and
the dedicated RunMetricsError have no remaining role under the DAG
executor: fx.evaluate (post-#445) now takes the user-facing metric
list directly, and the run_metrics-only "descriptive bundle" output
shape collapses into EvaluationResult.MetricResult. _project_factor
moves to factrix/_dag.py — the only surviving caller.

The _evaluate IC stage-1 batch-share test and the chunked / iter
tests are deleted alongside the helpers; they exercise the legacy
dispatch path that #449 (now folded into #448) tears down.

Refs #448.
