+++ mkdocs.yml
@@ -155,7 +155,6 @@ nav:
       - How-to:
           - Overview: guides/index.md
           - Cross-function reference: api/decision-tree.md
-          - Efficient data loading for large panels: guides/efficient-loading.md
           - Preparing data: guides/preparing-data.md
           - Information coefficient vs Fama-MacBeth: guides/choosing-metric.md
           - Panel vs timeseries: guides/panel-timeseries.md
@@ -181,7 +180,6 @@ nav:
           - Core:
               - AnalysisConfig: api/analysis-config.md
               - evaluate: api/evaluate.md
-              - run_metrics: api/run-metrics.md
           - Slicing:
               - by_slice: api/by-slice.md
               - slice_pairwise_test / slice_joint_test: api/slice-test.md
@@ -198,7 +196,6 @@ nav:
               - compare: api/compare.md
       - Results:
           - FactorProfile: api/factor-profile.md
-          - MetricsBundle: api/metrics-bundle.md
           - MetricOutput: api/metric-output.md
           - Identity / context: api/identity.md
           - Errors: api/errors.md
